refactor: 테트리스 게임 공통 부분 추상화 #62
